T2P migration
This patch will allow moving neutron from using the nsx_v3 plugin to the nsx_p plugin.
This includes:
- admin utility to move all resources to the policy api:
  nsxadmin -r nsx-migrate-t2p -o import (--verbose)
  This utility will:
  -- Migrate all neutron used & created resource using the nsx migration api
  -- roll back all resources in case it failed
  -- post migration fix some of the policy resources to better match the expectation
     of the policy plugin
- admin utility that will cleanup left overs in the nsx_v3 db:
  nsxadmin -r nsx-migrate-t2p -o clean-all
  (can be used, but everything should work without calling it as well)
- Some minor changes to the policy plugin and drivers to allow it to handle migrated resource
  which are a bit different than those created with the policy plugin
  -- Delete DHCP server config once a migrated network is deleted
  -- Update LB L7 rules by their name suffix as their full display name is unknown

+++ vmware_nsx/plugins/nsx_p/plugin.py
@@ -822,13 +822,19 @@ class NsxPolicyPlugin(nsx_plugin_common.NsxPluginV3Base):
         # MP MD proxy when this network is created.
         # If not - the port will not be found, and it is ok.
         # Note(asarfaty): In the future this code can be removed.
-        if not is_external_net and cfg.CONF.nsx_p.allow_passthrough:
+        # TODO(asarfaty): For migrated networks when the DB was not cleaned up
+        # This may actually delete a port the policy now control
+        if (not is_external_net and not is_nsx_net and
+            cfg.CONF.nsx_p.allow_passthrough):
             self._delete_nsx_port_by_network(network_id)
 
         # Delete the network segment from the backend
         if not is_external_net and not is_nsx_net:
             try:
                 self.nsxpolicy.segment.delete(network_id)
+                # In case of migrated network, a dhcp server config with
+                # the same id should also be deleted
+                self.nsxpolicy.dhcp_server_config.delete(network_id)
             except nsx_lib_exc.ResourceNotFound:
                 # If the resource was not found on the backend do not worry
                 # about it. The conditions has already been logged, so there
@@ -3921,7 +3927,7 @@ class NsxPolicyPlugin(nsx_plugin_common.NsxPluginV3Base):
                 rule_entry = self._create_security_group_backend_rule(
                     context, sg_id, rule, secgroup_logging,
                     is_provider_sg=is_provider_sg,
-                    create_related_resource=False)
+                    create_related_resource=True)
                 backend_rules.append(rule_entry)
 
             # Update the policy with all the rules.
